User can view users timeline by clicking username
Doing this was as simple as extracting the already present code blocks
(such as the code which generated messages and displayed them) and
extracting them to methods. Once the preset code was in methods it was
easy to manupulate the message streams.

I included underscore.js and used '_.flatten' on the array of all tweets, I
then used underscores '_.where' method to filter through them and return
only those tweets where the correct username property was set.

The only problem with this approach is the scripting takes up
considerable resources, which I checked out on the timeline tab on
chrome console. I don't think there is really any more efficient way to
do this though without a backend server. This runs in O(n) time.
Although there are many loops, none of them are entirely dependent on
each other. In actuality this may be n n2, n3, or even n4, but since we
never nests loops it never goes to the magnitude of n(^2). The reason
the scripting is slow is because of the size of our data set; our
browsers are not meant to process and render datasets of this size.